[CHORE/#28] BaseResponse 구현 및 적용
