+++ data/src/main/java/com/going/data/datasource/LoginDataSource.kt
@@ -1,11 +1,12 @@
 package com.going.data.datasource
 
+import com.going.data.dto.BaseResponse
 import com.going.data.dto.request.RequestLoginDto
-import com.going.data.dto.response.ResponseLoginDto
+import com.going.data.dto.response.LoginResponseDto
 
 interface LoginDataSource {
     suspend fun postLogin(
         Authorization: String,
         platform: RequestLoginDto,
-    ): ResponseLoginDto
+    ): BaseResponse<LoginResponseDto>
 }
