+++ data/src/main/java/com/going/data/dto/BaseResponse.kt
@@ -1,0 +1,14 @@
+package com.going.data.dto
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class BaseResponse<T>(
+    @SerialName("status")
+    val status: Int,
+    @SerialName("message")
+    val message: String,
+    @SerialName("data")
+    val data: T,
+)
