+++ data/src/main/java/com/going/data/service/LoginService.kt
@@ -1,7 +1,8 @@
 package com.going.data.service
 
+import com.going.data.dto.BaseResponse
 import com.going.data.dto.request.RequestLoginDto
-import com.going.data.dto.response.ResponseLoginDto
+import com.going.data.dto.response.LoginResponseDto
 import retrofit2.http.Body
 import retrofit2.http.Header
 import retrofit2.http.POST
@@ -11,5 +12,5 @@ interface LoginService {
     suspend fun postSignin(
         @Header("Authorization") Authorization: String,
         @Body body: RequestLoginDto,
-    ): ResponseLoginDto
+    ): BaseResponse<LoginResponseDto>
 }
